Simulate data based on item table

Once you've retrieved an item table using formr_items() you can use this function to sample data from the possible choices. At the moment random data is only generated for choice-type items and numeric ones, as these are most likely to enter data analysis. Does not yet handle dates, times, text, locations, colors

formr_simulate_from_items(item_list, n = 300)

Arguments

item_list
the result of a call to formr_connect()

n
defaults to 300

Examples

if (FALSE) {
formr_connect(email = 'you@example.net', password = 'zebrafinch' )
sim = formr_simulate_from_items(item_list = formr_items('training_diary'), n = 100)
summary(lm(pushups ~ pullups, data = sim))
}
items = formr_items(path = 
system.file('extdata/gods_example_items.json', package = 'formr', mustWork = TRUE))
fakedata = formr_simulate_from_items(items, n = 20)
fakedata[1:2,]

#>   id             created               ended            modified
#> 1  1 2021-03-27 13:31:03 2021-03-27 13:33:14 2021-03-27 13:33:14
#> 2  2 2021-04-05 00:19:17 2021-04-05 00:21:56 2021-04-05 00:21:56
#>                gods religiousness_1 religiousness_2R religiousness_3
#> 1 spaghetti_monster               4                1               3
#> 2 spaghetti_monster               1                3               3
#>   religiousness_4 prefer_1 prefer_2
#> 1               1        3        5
#> 2               2        2        4