Simulate data based on item table

Once you've retrieved an item table using formr_items() you can use this function to sample data from the possible choices. At the moment random data is only generated for choice-type items and numeric ones, as these are most likely to enter data analysis. Does not yet handle dates, times, text, locations, colors

formr_simulate_from_items(item_list, n = 300)

Arguments

item_list
the result of a call to formr_connect()

n
defaults to 300

Examples

# NOT RUN {
formr_connect(email = 'you@example.net', password = 'zebrafinch' )
sim = formr_simulate_from_items(item_list = formr_items('training_diary'), n = 100)
summary(lm(pushups ~ pullups, data = sim))
# }
items = formr_items(path =
system.file('extdata/gods_example_items.json', package = 'formr', mustWork = TRUE))
fakedata = formr_simulate_from_items(items, n = 20)
fakedata[1:2,]

#>   id             created               ended            modified    gods
#> 1  1 2017-11-22 13:31:03 2017-11-22 13:33:22 2017-11-22 13:33:22    glob
#> 2  2 2017-11-30 23:19:17 2017-11-30 23:21:42 2017-11-30 23:21:42 cthulhu
#>   religiousness_1 religiousness_2R religiousness_3 religiousness_4 prefer_1
#> 1               2                2               3               2        4
#> 2               4                3               2               1        2
#>   prefer_2
#> 1        1
#> 2        5