word_document from rmarkdown, but has an added option not to break on error

Exactly like rmarkdown::word_document(), but with one added argument

Usage

word_document(..., break_on_error = FALSE)

Arguments

...

all other arguments passed to rmarkdown::word_document()

break_on_error

should an error in the R code execution interrupt the rendering or should rendering continue, defaults to FALSE